Reorganized functional domain analysis
